fix: resolve 9 Rubocop offenses from CI
- Style/HashEachMethods: use each_key in base_handler
- Metrics/ParameterLists: combine observer params in movement_handler
- Metrics/MethodLength: extract collect_exit_descriptions helper
- Metrics/ClassLength: extract ContainerState concern from Game
- Style/Lambda: use lambda method for multiline scope in Message
- Layout/ExtraSpacing: fix spacing in multiplayer_system_test
- Rails/Pluck: use pluck in turn_manager_test
- Exclude test files from ParameterLists cop in rubocop config
