+++ app/models/game.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Game < ApplicationRecord
+  include ContainerState
+
   belongs_to :host, class_name: "User",
                     foreign_key: :created_by,
                     primary_key: :id,
@@ -119,66 +121,6 @@ class Game < ApplicationRecord
   def exit_revealed?(room_id, direction)
     exit_key = "#{room_id}_#{direction}"
     game_state.dig("revealed_exits", exit_key) || false
-  end
-
-  # Container state methods
-  def container_state(container_id)
-    game_state.dig("container_states", container_id.to_s)
-  end
-
-  def container_open?(container_id)
-    state = container_state(container_id)
-    return state["open"] if state
-
-    # If no state exists, check if container starts closed
-    item_def = world_snapshot.dig("items", container_id.to_s)
-    return true unless item_def&.dig("starts_closed")
-
-    false
-  end
-
-  def open_container(container_id)
-    self.game_state ||= {}
-    self.game_state["container_states"] ||= {}
-    self.game_state["container_states"][container_id.to_s] = { "open" => true }
-    save!
-  end
-
-  def close_container(container_id)
-    self.game_state ||= {}
-    self.game_state["container_states"] ||= {}
-    self.game_state["container_states"][container_id.to_s] = { "open" => false }
-    save!
-  end
-
-  def container_contents(container_id)
-    # Get original contents from world snapshot
-    original_contents = world_snapshot.dig("items", container_id.to_s, "contents") || []
-
-    # Get removed items from game state
-    removed_items = game_state.dig("container_states", container_id.to_s, "removed_items") || []
-
-    # Return contents minus removed items
-    original_contents - removed_items
-  end
-
-  def remove_from_container(container_id, item_id)
-    self.game_state ||= {}
-    self.game_state["container_states"] ||= {}
-    self.game_state["container_states"][container_id.to_s] ||= {}
-    self.game_state["container_states"][container_id.to_s]["removed_items"] ||= []
-    self.game_state["container_states"][container_id.to_s]["removed_items"] << item_id
-    self.game_state["container_states"][container_id.to_s]["removed_items"].uniq!
-    save!
-  end
-
-  def add_to_container(container_id, item_id)
-    self.game_state ||= {}
-    self.game_state["container_states"] ||= {}
-    self.game_state["container_states"][container_id.to_s] ||= {}
-    self.game_state["container_states"][container_id.to_s]["removed_items"] ||= []
-    self.game_state["container_states"][container_id.to_s]["removed_items"].delete(item_id)
-    save!
   end
 
   def turn_count
